Make record to model logic shared (#202)
* [WIP] [skip-ci] Make record to model logic shared

Presentation Record logic shared fully


* [WIP] [skip ci] Added shared logic for connections hook and connections


* feat: shared models and shared conversion of models

between webhooks and cloudapi
