+++ app/generic/issuer/facades/acapy_issuer_v1.py
@@ -17,7 +17,10 @@ from app.generic.issuer.models import (
     Credential,
 )
 from app.generic.issuer.facades.acapy_issuer_utils import cred_id_no_version
-from shared_models import CredentialExchange, IssueCredentialProtocolVersion
+from shared_models import (
+    CredentialExchange,
+    credential_record_to_model_v1,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -105,36 +108,7 @@ class IssuerV1(Issuer):
 
     @classmethod
     def __record_to_model(cls, record: V10CredentialExchange) -> CredentialExchange:
-        attributes = cls.__attributes_from_record(record)
-
-        return CredentialExchange(
-            credential_id=f"v1-{record.credential_exchange_id}",
-            role=record.role,
-            created_at=record.created_at,
-            updated_at=record.updated_at,
-            attributes=attributes,
-            protocol_version=IssueCredentialProtocolVersion.v1,
-            schema_id=record.schema_id,
-            credential_definition_id=record.credential_definition_id,
-            state=cls.__v1_state_to_rfc_state(record.state),
-            connection_id=record.connection_id,
-        )
-
-    @classmethod
-    def __attributes_from_record(
-        cls, record: V10CredentialExchange
-    ) -> Optional[Dict[str, str]]:
-        preview = None
-
-        if (
-            record.credential_proposal_dict
-            and record.credential_proposal_dict.credential_proposal
-        ):
-            preview = record.credential_proposal_dict.credential_proposal
-
-        return (
-            {attr.name: attr.value for attr in preview.attributes} if preview else None
-        )
+        return credential_record_to_model_v1(record)
 
     @classmethod
     def __preview_from_attributes(
@@ -147,22 +121,3 @@ class IssuerV1(Issuer):
                 for name, value in attributes.items()
             ]
         )
-
-    @classmethod
-    def __v1_state_to_rfc_state(cls, state: Optional[str]) -> Optional[str]:
-        translation_dict = {
-            "proposal_sent": "proposal-sent",
-            "proposal_received": "proposal-received",
-            "offer_sent": "offer-sent",
-            "offer_received": "offer-received",
-            "request_sent": "request-sent",
-            "request_received": "request-received",
-            "credential_issued": "credential-issued",
-            "credential_received": "credential-received",
-            "credential_acked": "done",
-        }
-
-        if not state or state not in translation_dict:
-            return None
-
-        return translation_dict[state]
