+++ app/generic/verifier/facades/acapy_verifier_v1.py
@@ -14,7 +14,6 @@ from aries_cloudcontroller.model.v10_presentation_exchange import (
 from fastapi.exceptions import HTTPException
 from pydantic.typing import NoneType
 
-import app.generic.verifier.facades.acapy_verifier_utils as utils
 from app.generic.verifier.facades.acapy_verifier import Verifier
 from app.generic.verifier.models import (
     AcceptProofRequest,
@@ -22,7 +21,11 @@ from app.generic.verifier.models import (
     RejectProofRequest,
     SendProofRequest,
 )
-from shared_models import PresentationExchange
+from shared_models import (
+    PresentationExchange,
+    presentation_record_to_model as record_to_model,
+    pres_id_no_version,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -43,12 +46,12 @@ class VerifierV1(Verifier):
                 )
             )
         )
-        return utils.record_to_model(presentation_exchange)
+        return record_to_model(presentation_exchange)
 
     @classmethod
     async def get_credentials_for_request(cls, controller: AcaPyClient, proof_id: str):
         try:
-            pres_ex_id = utils.pres_id_no_version(proof_id=proof_id)
+            pres_ex_id = pres_id_no_version(proof_id=proof_id)
             return await controller.present_proof_v1_0.get_matching_credentials(
                 pres_ex_id=pres_ex_id
             )
@@ -60,7 +63,7 @@ class VerifierV1(Verifier):
     async def get_proof_records(cls, controller: AcaPyClient):
         try:
             presentation_exchange = await controller.present_proof_v1_0.get_records()
-            return [utils.record_to_model(rec) for rec in presentation_exchange.results]
+            return [record_to_model(rec) for rec in presentation_exchange.results]
         except Exception as e:
             logger.error(f"{e!r}")
             raise e from e
@@ -68,11 +71,11 @@ class VerifierV1(Verifier):
     @classmethod
     async def get_proof_record(cls, controller: AcaPyClient, proof_id: str):
         try:
-            pres_ex_id = utils.pres_id_no_version(proof_id)
+            pres_ex_id = pres_id_no_version(proof_id)
             presentation_exchange = await controller.present_proof_v1_0.get_record(
                 pres_ex_id=pres_ex_id
             )
-            return utils.record_to_model(presentation_exchange)
+            return record_to_model(presentation_exchange)
         except Exception as e:
             logger.error(f"{e!r}")
             raise e from e
@@ -80,7 +83,7 @@ class VerifierV1(Verifier):
     @classmethod
     async def delete_proof(cls, controller: AcaPyClient, proof_id: str):
         try:
-            pres_ex_id = utils.pres_id_no_version(proof_id=proof_id)
+            pres_ex_id = pres_id_no_version(proof_id=proof_id)
             return await controller.present_proof_v1_0.delete_record(
                 pres_ex_id=pres_ex_id
             )
@@ -108,7 +111,7 @@ class VerifierV1(Verifier):
                 )
             else:
                 raise NotImplementedError
-            return utils.record_to_model(presentation_exchange)
+            return record_to_model(presentation_exchange)
         except (AttributeError, TypeError) as e:
             raise NotImplementedError(f"{e!r}")
         except Exception as e:
@@ -118,18 +121,18 @@ class VerifierV1(Verifier):
     async def accept_proof_request(
         cls, controller: AcaPyClient, proof_request: AcceptProofRequest
     ) -> PresentationExchange:
-        proof_id = utils.pres_id_no_version(proof_id=proof_request.proof_id)
+        proof_id = pres_id_no_version(proof_id=proof_request.proof_id)
         presentation_record = await controller.present_proof_v1_0.send_presentation(
             pres_ex_id=proof_id, body=proof_request.presentation_spec
         )
-        return utils.record_to_model(presentation_record)
+        return record_to_model(presentation_record)
 
     @classmethod
     async def reject_proof_request(
         cls, controller: AcaPyClient, proof_request: RejectProofRequest
     ) -> None:
         # get the record
-        proof_id = utils.pres_id_no_version(proof_id=proof_request.proof_id)
+        proof_id = pres_id_no_version(proof_id=proof_request.proof_id)
         proof_request_res = await controller.present_proof_v1_0.get_record(
             pres_ex_id=proof_id
         )
