+++ shared_models/__init__.py
@@ -1,9 +1,15 @@
 from .shared_models import (
     TopicItem,
     Connection,
+    ConnectionsHook,
+    credential_record_to_model_v1,
+    credential_record_to_model_v2,
+    conn_record_to_connection,
     PresentationExchange,
     HookBase,
     ProofRequestProtocolVersion,
     CredentialExchange,
     IssueCredentialProtocolVersion,
+    presentation_record_to_model,
+    pres_id_no_version,
 )
